+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 
 - Fixed NullPointerException for `onSinglePointerTouchEvent` of null key
+- Fixed key press colour change dependent on activity background
 - Reduced space bar height
 - Reduced digit row key height
 - Added fillColour attribute for the keyboard as a whole